chore(pipeline): sources/src/processors/sources -> sources/
