+++ src/query/pipeline/sources/src/empty_source.rs
