+++ src/query/pipeline/sources/src/input_formats/impls/input_format_csv.rs
@@ -34,15 +34,15 @@ use common_meta_types::OnErrorMode;
 use common_meta_types::StageFileFormatType;
 use csv_core::ReadRecordResult;
 
-use crate::processors::sources::input_formats::impls::input_format_tsv::format_column_error;
-use crate::processors::sources::input_formats::input_format_text::AligningStateCommon;
-use crate::processors::sources::input_formats::input_format_text::AligningStateTextBased;
-use crate::processors::sources::input_formats::input_format_text::BlockBuilder;
-use crate::processors::sources::input_formats::input_format_text::InputFormatTextBase;
-use crate::processors::sources::input_formats::input_format_text::RowBatch;
-use crate::processors::sources::input_formats::InputContext;
-use crate::processors::sources::input_formats::InputError;
-use crate::processors::sources::input_formats::SplitInfo;
+use crate::input_formats::impls::input_format_tsv::format_column_error;
+use crate::input_formats::input_format_text::AligningStateCommon;
+use crate::input_formats::input_format_text::AligningStateTextBased;
+use crate::input_formats::input_format_text::BlockBuilder;
+use crate::input_formats::input_format_text::InputFormatTextBase;
+use crate::input_formats::input_format_text::RowBatch;
+use crate::input_formats::InputContext;
+use crate::input_formats::InputError;
+use crate::input_formats::SplitInfo;
 
 pub struct InputFormatCSV {}
