+++ src/query/pipeline/sources/src/input_formats/impls/input_format_tsv.rs
@@ -30,11 +30,11 @@ use common_io::format_diagnostic::verbose_string;
 use common_meta_types::OnErrorMode;
 use common_meta_types::StageFileFormatType;
 
-use crate::processors::sources::input_formats::input_format_text::AligningStateRowDelimiter;
-use crate::processors::sources::input_formats::input_format_text::BlockBuilder;
-use crate::processors::sources::input_formats::input_format_text::InputFormatTextBase;
-use crate::processors::sources::input_formats::input_format_text::RowBatch;
-use crate::processors::sources::input_formats::InputError;
+use crate::input_formats::input_format_text::AligningStateRowDelimiter;
+use crate::input_formats::input_format_text::BlockBuilder;
+use crate::input_formats::input_format_text::InputFormatTextBase;
+use crate::input_formats::input_format_text::RowBatch;
+use crate::input_formats::InputError;
 
 pub struct InputFormatTSV {}
