+++ src/query/pipeline/sources/src/input_formats/input_pipeline.rs
@@ -31,14 +31,14 @@ use futures_util::AsyncReadExt;
 use futures_util::StreamExt;
 use opendal::raw::CompressAlgorithm;
 
-use crate::processors::sources::input_formats::beyond_end_reader::BeyondEndReader;
-use crate::processors::sources::input_formats::input_context::InputContext;
-use crate::processors::sources::input_formats::input_context::InputPlan;
-use crate::processors::sources::input_formats::input_context::StreamPlan;
-use crate::processors::sources::input_formats::input_split::SplitInfo;
-use crate::processors::sources::input_formats::source_aligner::Aligner;
-use crate::processors::sources::input_formats::source_deserializer::DeserializeSource;
-use crate::processors::sources::input_formats::transform_deserializer::DeserializeTransformer;
+use crate::input_formats::beyond_end_reader::BeyondEndReader;
+use crate::input_formats::input_context::InputPlan;
+use crate::input_formats::input_context::StreamPlan;
+use crate::input_formats::source_aligner::Aligner;
+use crate::input_formats::source_deserializer::DeserializeSource;
+use crate::input_formats::transform_deserializer::DeserializeTransformer;
+use crate::input_formats::InputContext;
+use crate::input_formats::SplitInfo;
 
 pub struct Split<I: InputFormatPipe> {
     pub(crate) info: Arc<SplitInfo>,
