+++ src/query/pipeline/sources/src/one_block_source.rs
