+++ src/query/pipeline/sources/tests/it/main.rs
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod processors;
+mod sources;
