+++ src/query/service/src/interpreters/interpreter_insert_v2.rs
@@ -55,8 +55,8 @@ use common_io::cursor_ext::ReadCheckPointExt;
 use common_meta_types::FileFormatOptions;
 use common_meta_types::UserStageInfo;
 use common_pipeline_core::Pipeline;
-use common_pipeline_sources::processors::sources::AsyncSource;
-use common_pipeline_sources::processors::sources::AsyncSourcer;
+use common_pipeline_sources::AsyncSource;
+use common_pipeline_sources::AsyncSourcer;
 use common_pipeline_transforms::processors::transforms::Transform;
 use common_sql::evaluator::BlockOperator;
 use common_sql::evaluator::CompoundBlockOperator;
