+++ src/query/service/src/pipelines/processors/mod.rs
@@ -15,14 +15,6 @@
 pub use common_pipeline_core::processors::*;
 pub(crate) mod transforms;
 
-use common_pipeline_sources::processors::sources;
-pub use sources::AsyncSource;
-pub use sources::AsyncSourcer;
-pub use sources::BlocksSource;
-pub use sources::EmptySource;
-pub use sources::StreamSource;
-pub use sources::SyncSource;
-pub use sources::SyncSourcer;
 pub use transforms::AggregatorParams;
 pub use transforms::AggregatorTransformParams;
 pub use transforms::BlockCompactor;
