+++ src/query/service/src/servers/http/clickhouse_handler.rs
@@ -29,8 +29,8 @@ use common_expression::TableSchemaRef;
 use common_formats::ClickhouseFormatType;
 use common_formats::FileFormatOptionsExt;
 use common_formats::FileFormatTypeExt;
-use common_pipeline_sources::processors::sources::input_formats::InputContext;
-use common_pipeline_sources::processors::sources::input_formats::StreamingReadBatch;
+use common_pipeline_sources::input_formats::InputContext;
+use common_pipeline_sources::input_formats::StreamingReadBatch;
 use common_sql::plans::InsertInputSource;
 use common_sql::plans::Plan;
 use common_sql::Planner;
