+++ src/query/service/src/servers/http/v1/load.rs
@@ -24,8 +24,8 @@ use common_base::runtime::TrySpawn;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::infer_table_schema;
-use common_pipeline_sources::processors::sources::input_formats::InputContext;
-use common_pipeline_sources::processors::sources::input_formats::StreamingReadBatch;
+use common_pipeline_sources::input_formats::InputContext;
+use common_pipeline_sources::input_formats::StreamingReadBatch;
 use common_sql::plans::InsertInputSource;
 use common_sql::plans::Plan;
 use common_sql::Planner;
