+++ src/query/service/src/table_functions/sync_crash_me.rs
@@ -33,12 +33,12 @@ use common_expression::TableSchema;
 use common_meta_app::schema::TableIdent;
 use common_meta_app::schema::TableInfo;
 use common_meta_app::schema::TableMeta;
+use common_pipeline_sources::SyncSource;
+use common_pipeline_sources::SyncSourcer;
 use futures::Stream;
 
 use crate::pipelines::processors::port::OutputPort;
 use crate::pipelines::processors::processor::ProcessorPtr;
-use crate::pipelines::processors::SyncSource;
-use crate::pipelines::processors::SyncSourcer;
 use crate::pipelines::Pipeline;
 use crate::sessions::TableContext;
 use crate::storages::Table;
