+++ src/query/service/tests/it/storages/fuse/table_test_fixture.rs
@@ -40,6 +40,7 @@ use common_expression::TableSchemaRefExt;
 use common_meta_app::schema::DatabaseMeta;
 use common_meta_types::StorageFsConfig;
 use common_meta_types::StorageParams;
+use common_pipeline_sources::BlocksSource;
 use common_sql::plans::CreateDatabasePlan;
 use common_sql::plans::CreateTablePlanV2;
 use common_storages_fuse::FuseTable;
@@ -50,7 +51,6 @@ use databend_query::interpreters::Interpreter;
 use databend_query::interpreters::InterpreterFactory;
 use databend_query::pipelines::executor::ExecutorSettings;
 use databend_query::pipelines::executor::PipelineCompleteExecutor;
-use databend_query::pipelines::processors::BlocksSource;
 use databend_query::pipelines::PipelineBuildResult;
 use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
