+++ src/query/storages/memory/src/memory_table.rs
@@ -39,8 +39,8 @@ use common_pipeline_core::processors::port::OutputPort;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_core::Pipeline;
 use common_pipeline_sinks::ContextSink;
-use common_pipeline_sources::processors::sources::SyncSource;
-use common_pipeline_sources::processors::sources::SyncSourcer;
+use common_pipeline_sources::SyncSource;
+use common_pipeline_sources::SyncSourcer;
 use common_storage::StorageMetrics;
 use once_cell::sync::Lazy;
 use parking_lot::Mutex;
