+++ src/query/storages/stage/src/stage_table.rs
@@ -38,8 +38,8 @@ use common_expression::DataBlock;
 use common_meta_app::schema::TableInfo;
 use common_meta_types::UserStageInfo;
 use common_pipeline_core::Pipeline;
-use common_pipeline_sources::processors::sources::input_formats::InputContext;
-use common_pipeline_sources::processors::sources::input_formats::SplitInfo;
+use common_pipeline_sources::input_formats::InputContext;
+use common_pipeline_sources::input_formats::SplitInfo;
 use common_storage::init_stage_operator;
 use opendal::Operator;
 use parking_lot::Mutex;
